requirements.txt
# Core dependencies
openai-whisper>=20231117      # 音声認識
Pillow>=10.0.0                # 画像処理
python-dotenv>=1.0.0          # 環境変数
PyYAML>=6.0.1                 # 設定ファイル
click>=8.1.0                  # CLI
tqdm>=4.66.0                  # プログレスバー
colorama>=0.4.6               # カラー出力

# Video processing
ffmpeg-python>=0.2.0          # 動画処理

# Text processing
beautifulsoup4>=4.12.0        # HTML処理
markdown>=3.5.0               # Markdown変換
python-slugify>=8.0.0         # URLスラッグ生成

# Optional for enhanced features
requests>=2.31.0              # Web API
jinja2>=3.1.0                 # テンプレートエンジン